UPDATE: if a dropdown menu has entries with urls with fragments, force the menu open if the control is active
If we have urls with fragments in our menu, it is likely that the submenu is a local table of contents for the page.  Since fragments are not passed from the web browser to the server, we won't know if someone had clicked on one of those items previously to navigate to where they are, so  let's just open the menu so users can see their choices.